refactor: observer in MessageSeen.jsx (#669)
