+++ client/src/components/Chat/MessageSeen.jsx
@@ -3,18 +3,20 @@ import PropTypes from 'prop-types';
 
 import useIsTabActive from 'src/hooks/useIsTabActive';
 
-import useChatUtils from 'src/lib/chatSocket';
+import useObserver from 'src/hooks/useObserver';
+
 import { useApp } from 'src/context/AppContext';
 import { useChat } from 'src/context/ChatContext';
-import { socket } from 'src/lib/socketConnection';
 
 const MessageSeen = ({ isRead, isSender }) => {
 	const { app } = useApp();
 
 	const isTabVisible = useIsTabActive();
 
-	const { messages: state, receiveMessage } = useChat();
-	const { seenMessage } = useChatUtils(socket);
+	const observer = useObserver(isSender);
+
+	const { messages: state } = useChat();
+	
 
 	const sortedMessages = useMemo(
 		() =>
@@ -28,27 +30,7 @@ const MessageSeen = ({ isRead, isSender }) => {
 
 	useEffect(() => {
 		// Initialize Intersection Observer
-		const observer = new IntersectionObserver(
-			(entries) => {
-				entries.forEach((entry) => {
-					if (entry.isIntersecting && !isSender) {
-						// Mark the message as read
-						const messageId = entry.target.getAttribute('id').split('-')[1];
-						try {
-							seenMessage({
-								messageId,
-								chatId: app.currentChatId,
-							});
-						} catch (e) {
-							return;
-						}
-						receiveMessage(messageId, app.currentChatId);
-					}
-				});
-			},
-			{ threshold: 0.5 }
-			// Trigger when 50% of the element is in the viewport
-		);
+		observer.connect();
 
 		sortedMessages.forEach((message) => {
 			if (message.isRead) {
